auth modals refactor
