+++ src/lib/components/Modals/LoginModal.tsx
@@ -17,11 +17,8 @@ import {
 import Visibility from '@material-ui/icons/Visibility';
 import VisibilityOff from '@material-ui/icons/VisibilityOff';
 import { RegistrationModal } from './RegistrationModal';
-
-export interface IInputs {
-  email: string;
-  password: string;
-}
+import { emailValidationObj, passwordValidationObj } from './validationRules';
+import { IInputs } from '../../types';
 
 export const LoginModal: React.FC = () => {
   const [loginOpen, setLoginOpen] = useState(false);
@@ -66,32 +63,6 @@ export const LoginModal: React.FC = () => {
         <span style={{ color: 'red' }}>{errorsObj[inputName].message}</span>
       )
     );
-  };
-
-  const emailValidationObj = {
-    required: {
-      value: true,
-      message: "Це поле є обов'язковим",
-    },
-    pattern: {
-      value: /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i,
-      message: 'Неправильний формат email',
-    },
-  };
-
-  const passwordValidationObj = {
-    required: {
-      value: true,
-      message: "Це поле є обов'язковим",
-    },
-    minLength: {
-      value: 4,
-      message: 'Пароль повинен містити щонайменше 4 символи',
-    },
-    maxLength: {
-      value: 16,
-      message: 'Пароль повинен містити щонайбільше 16 символи',
-    },
   };
 
   return (
